packages/sdk-python/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pocketping"
version = "0.1.0"
description = "Python SDK for PocketPing - real-time customer chat with mobile notifications"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
authors = [
    { name = "PocketPing Contributors" }
]
keywords = ["chat", "support", "telegram", "discord", "slack", "notifications", "fastapi"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "pydantic>=2.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
fastapi = [
    "fastapi>=0.100.0",
    "websockets>=12.0",
]
telegram = [
    "python-telegram-bot>=20.0",
]
discord = [
    "discord.py>=2.3.0",
]
slack = [
    "slack-sdk>=3.23.0",
]
ai = [
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "anthropic>=0.18.0",
]
all = [
    "pocketping[fastapi,telegram,discord,slack,ai]",
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.25.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/Ruwad-io/pocketping"
Documentation = "https://github.com/Ruwad-io/pocketping/tree/main/packages/sdk-python#readme"
Repository = "https://github.com/Ruwad-io/pocketping"

[tool.hatch.build.targets.wheel]
packages = ["src/pocketping"]

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pytest.ini_options]
asyncio_mode = "auto"
